+++ tests/testthat/test_column.R
@@ -38,19 +38,22 @@ test_that("posixColumnAtPosition() and firstPosixColumn() work", {
 
 test_that("moveColumnsToFront() works", {
   
-  x <- data.frame(a = 1:5, b = 2:6, c = 3:7)
+  df <- data.frame(a = 1:5, b = 2:6, c = 3:7)
+  df_one_column <- data.frame(a = 1:4)
+  
+  y_1 <- moveColumnsToFront(df, "b")
+  y_2 <- moveColumnsToFront(df, c("b", "a"))
+  y_3 <- moveColumnsToFront(df)
 
-  y_1 <- moveColumnsToFront(x, "b")
-  y_2 <- moveColumnsToFront(x, c("b", "a"))
-  y_3 <- moveColumnsToFront(x)
-
-  expect_identical(dim(x), dim(y_1))
-  expect_identical(dim(x), dim(y_2))  
+  expect_identical(dim(df), dim(y_1))
+  expect_identical(dim(df), dim(y_2))  
 
   expect_identical(names(y_1)[1], "b")
   expect_identical(names(y_2)[1:2], c("b", "a"))
   
-  expect_identical(y_3, x)
+  expect_identical(y_3, df)
+  
+  expect_identical(moveColumnsToFront(df_one_column, "a"), df_one_column)
 })
 
 test_that("checkForMissingColumns() works", {